feat: check if subs exist

+++ code/src/app/service/subscription/delete-user-subscription.service.ts
@@ -1,7 +1,10 @@
 import { DeleteUserSubscriptionDTO } from '@/dtos/subscription/delete-subscroption.dto';
 import httpError from 'http-errors';
 import { IFindEventByIdRepository } from '@/interfaces/event/event-repository.interface';
-import { IDeleteUserSubscriptionRepository } from '@/interfaces/subscription/subscription-repository.interface';
+import {
+  IDeleteUserSubscriptionRepository,
+  IFindSubscriptionByUserIdAndEventIdRepository,
+} from '@/interfaces/subscription/subscription-repository.interface';
 import { IDeleteUserSubscriptionService } from '@/interfaces/subscription/subscription-service.interface';
 import { deleteUserSubscription } from '@/schemas/subscription.schema';
 export class DeleteUserSubscriptionService
@@ -9,7 +12,8 @@ export class DeleteUserSubscriptionService
 {
   constructor(
     private deleteUserSubscriptionRepository: IDeleteUserSubscriptionRepository,
-    private findEventByIdRepository: IFindEventByIdRepository
+    private findEventByIdRepository: IFindEventByIdRepository,
+    private findSubscriptionByUserIdAndEventIdRepository: IFindSubscriptionByUserIdAndEventIdRepository
   ) {}
 
   async execute(params: DeleteUserSubscriptionDTO): Promise<void> {
@@ -17,8 +21,22 @@ export class DeleteUserSubscriptionService
 
     const event = await this.findEventByIdRepository.findById(eventId);
 
+    console.log(event);
+
     if (!event) {
       throw httpError.NotFound('Event not found');
+    }
+
+    const subscription =
+      await this.findSubscriptionByUserIdAndEventIdRepository.findByUserIdAndEventId(
+        {
+          userId,
+          eventId,
+        }
+      );
+
+    if (!subscription) {
+      throw httpError.NotFound('Subscription not found');
     }
 
     await this.deleteUserSubscriptionRepository.delete({ userId, eventId });
